+++ R/create_transposed_data.R
@@ -22,6 +22,8 @@
 #'
 #'
 #' library(dplyr)
+#' library(magrittr)
+#' library(lazytrade)
 #'
 #' # usind a sample data
 #' data(indicator_dataset)