+++ NEWS.md
@@ -2,12 +2,32 @@
 
 ## Planned Changes
 
-* Add folder generation function
-* Put option predict ahead as a parameter (now 34 bars)
-* aml_make_model use parameter to split data with specific n of rows e.g. n_rows
 * cleanup of deprecated functions
-* create option to set fixed structure of the neural network
-* make a major release
+* setup github actions
+
+# lazytrade 0.5.1
+
+# Version 0.5.1
+
+## Changes
+
+### New Function `aml_simulation`
+
+* designed to test different inputs
+
+### Function `aml_make_model`
+
+* fix bug in not allowing use of full input to the model
+* create new option with fix neural network structure
+* add parameter to split data with specific n of rows
+* add parameter num_bars_ahead, default is 34 bars
+* add options for deep learning function parameters search, `num_epochs`, `fixed_nn_struct`
+* add option `objective_test` to allow strategy test inside the function to select best model based on specific objective
+
+### Function `aml_collect_data`
+
+* more accurate calculation of Label column using tick data
+* added fail safe scenarios
 
 # lazytrade 0.4.5